added slow dom tree, basic tests

+++ ir/typedefs.h
@@ -1,7 +1,7 @@
 #ifndef ___TYPEDEFS_H_INCLUDED___
 #define ___TYPEDEFS_H_INCLUDED___
 
-typedef int IdType;
+typedef size_t IdType;
 typedef int ImmType;
 typedef uint ArgNumType;
